Get terrestrial carbon to balance right
UNIT GtC per PCal I assume is for crops (corn, wheat, soybean, yada yada yada) because that is where it was used initially. I also assume that value was pulled from LPJML for those species. My gut says that number shouldn't be the same for grass in a field that is being consumed right? So that is what the GtC per Pcal of grazing multipler is taking into account. I set it to 5% to reproduce what @jnnsbrr was wanting for Carbon from grazing... I have no idea if that is right or wrong.

I added a new component to "yield" (grass net primary production) of grassland... which is the amount of grassland itself. If grassland goes down, that means everything else being equal the yield of the average Hectacre of land should go up! THe assumption being the worst most unproductive land is abandoned and when you abandon bad land, the average for all land goes up!

I recalibrated changes in land use types to nail the two land use change based balances re: deforestation and young forests.

I added a new calibration to set the NPP parameters in the grass and forest modules to get the terrestial carbon balance in line.